Add navigation buttons and refactor UI components
- Added new styles for "Home", "Left", and "Right" buttons in `App.xaml`.
- Included new image resources (`house-icon.png`, `line-angle-left-icon.png`, `line-angle-right-icon.png`) in `DocumentVersionControlSystem.UI.csproj`.
- Updated `MainWindow.xaml` to include new navigation buttons ("Home", "Previous", "Next") and reorganized the layout.
- Refactored `MainWindow.xaml.cs` to remove unused methods and variables, add navigation logic, integrate `HomePage`, and adjust grid layout logic.
- Simplified `DocumentViewerPage.xaml` by removing unnecessary layout elements and keeping essential controls.
- Updated event handlers and navigation logic to support new button functionalities and page transitions.
- Refactored `DocumentViewerPage.xaml.cs` to remove unused `using` directives and redundant methods, remove `DocumentManager` dependency, and update `ReadDocument` method.
- Added `HomePage.xaml` and `HomePage.xaml.cs` to define and implement a new `HomePage` with a `ScrollViewer` and `Grid` layout.
- Simplified `VersionDetailsPage.xaml` by removing `Border` and `GridSplitter` elements, and directly including a `Button` and `TextBox`.
- Refactored `VersionDetailsPage.xaml.cs` to remove unused `using` directives and redundant methods, and update `RefreshWindow` method.
