+++ DocumentVersionControlSystem.UI/Windows/HomePage.xaml
@@ -1,0 +1,29 @@
+﻿<Page x:Class="DocumentVersionControlSystem.UI.Windows.HomePage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:DocumentVersionControlSystem.UI.Windows"
+      mc:Ignorable="d" 
+      d:DesignHeight="515" d:DesignWidth="850"
+      Title="HomePage" Loaded="Page_Loaded">
+
+    <ScrollViewer>
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="203" />
+                <ColumnDefinition Width="*" />
+            </Grid.ColumnDefinitions>
+
+            <Grid.RowDefinitions>
+                <RowDefinition Height="32" />
+                <RowDefinition Height="*" />
+            </Grid.RowDefinitions>
+
+            <!-- Кнопки у правій колонці -->
+            <StackPanel Name="DocumentStackPanel" Grid.Column="1" Grid.Row="1">
+                <Grid x:Name="ButtonGrid"/>
+            </StackPanel>
+        </Grid>
+    </ScrollViewer>
+</Page>
