+++ DocumentVersionControlSystem.UI/Windows/VersionDetailsPage.xaml
@@ -5,39 +5,24 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:DocumentVersionControlSystem.UI.Windows"
         mc:Ignorable="d"
-        Title="MainWindow" Height="515" Width="850" Background="#D9D9D9">
-    <Border BorderBrush="Black" BorderThickness="1">
-        <Grid>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="203" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
+      d:DesignHeight="515" d:DesignWidth="850"
+        Title="MainWindow">
 
-            <Grid.RowDefinitions>
-                <RowDefinition Height="32" />
-                <RowDefinition Height="*" />
-            </Grid.RowDefinitions>
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="203" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
 
-            <TextBlock Text="Document Versions" Grid.Column="0" FontSize="16" TextAlignment="Center" VerticalAlignment="Center" Height="29" Width="203"/>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="32" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        
+        <Button Grid.Column="1" Content="Apply version" Margin="0,0,10,8" Style="{StaticResource CustomButtonStyle}" Width="100" Height="20" HorizontalAlignment="Right" FontSize="12" VerticalAlignment="Bottom" Click="ApplyVersionButton_Click" />
+        <TextBox x:Name="TextBox" IsReadOnly="True" Grid.Column="1" Grid.Row="1" TextWrapping="Wrap" Margin="10,10,10,80" BorderBrush="Black" BorderThickness="1.5" Style="{StaticResource TextBoxStyle}"/>
 
-            <TextBlock Text="Document Version Control System" Grid.Column="1" Grid.Row="0" FontSize="16" TextAlignment="Center" VerticalAlignment="Bottom" Height="29" Width="545"/>
+        <TextBlock x:Name="TextBlock" Grid.Column="1" Grid.Row="1" TextWrapping="Wrap" Margin="10,0,0,50" VerticalAlignment="Bottom"/>
 
-            <GridSplitter Grid.Column="0" Grid.RowSpan="2" Width="3" Background="Black"
-              HorizontalAlignment="Stretch" Margin="200,0,0,0" Cursor="Arrow"/>
-
-            <GridSplitter Grid.Row="0" Grid.ColumnSpan="2" Height="3" Background="Black"
-              HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="0,28,0,0" Cursor="Arrow" />
-
-            <!-- Кнопки у лівій колонці -->
-            <ScrollViewer Grid.Column="0" Grid.Row="1" VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Disabled">
-                <StackPanel x:Name="ButtonStackPanel" Grid.Column="0" Grid.Row="1"/>
-            </ScrollViewer>
-
-            <TextBox x:Name="TextBox" IsReadOnly="True" Grid.Column="1" Grid.Row="1" TextWrapping="Wrap" Margin="10,10,10,80" BorderBrush="Black" BorderThickness="1.5" Style="{StaticResource TextBoxStyle}"/>
-
-            <TextBlock x:Name="TextBlock" Grid.Column="1" Grid.Row="1" TextWrapping="Wrap" Margin="10,0,0,50" VerticalAlignment="Bottom"/>
-            <Button Grid.Column="1" Content="Apply version" Margin="0,0,10,8" Style="{StaticResource CustomButtonStyle}" Width="100" Height="20" HorizontalAlignment="Right" FontSize="12" VerticalAlignment="Bottom" Click="ApplyVersionButton_Click" />
-
-        </Grid>
-    </Border>
+    </Grid>
 </Page>
